Re-organize cropduster unit tests
